Improve benchmark reporting
When benchmarking a synchronization primitive, raw throughput is not the
only statistic that matters. For example, it is also important that
threads do not starve waiting to acquire a lock.

Improve the benchmark reporting so that it reports the throughput per
thread per second.  This makes it possible to compare benchmarks across
different numbers of threads.  Without this, benchmark runs with more
threads would naturally appear to have higher throughput.  Additionally,
include the median and standard deviation for the throughput per thread
per second so that starvation issues become more apparent.

Signed-off-by: Chirantan Ekbote <chirantan@google.com>