Add a for_lock method on RwLockReadGuard and RwLockWriteGuard.
In Servo, I’m trying a scheme where mupltiple objects, each in Arc<_>,
are protected by a single shared RwLock.
This lock can be aquired once to access many objects.

https://bugzilla.mozilla.org/show_bug.cgi?id=1335941#c10

On each access, we assert that the guard being used indeed is for
the same lock that protects the object being accessed.
This is done by comparing `*const RwLock<()>` pointers.
I currently keep this pointer in my guard wrapper types,
but this is wasted space since this crate’s guards already
hold a reference to the lock.

This commit adds public API to guards to expose this reference.