lock_api: Make RawMutex::unlock() unsafe.
Unlocking a Mutex that you do not hold is likely to cause undefined behavior. If unlock()
is a safe method, then it likely impossible to provide a sound and efficient implemention
of the method. Since this API is usually only called by the lock_api Mutex structs, this
change should be invisible to the users, but clarifies what guarantees implementers must make.