Bump MSRV to 1.59
In PR #309 [1], usage of the `asm` feature was promoted because it is now
part of the stable Rust channel.

This, however, causes build failures when using any Rust version between
1.49 (the documented minimum version) up to 1.58, because the `asm`
feature has only become stable in Rust 1.59 [2].

Bump the documented minimum supported Rust version (MSRV) to 1.59.

Also, add a `rust_version` entry to `Cargo.toml` in order to codify
the MSRV for `cargo` to pick it up. This leads to a more helpful and
more actionable error message, for example:

> error: package `parking_lot v0.12.1 ([…])` cannot be built because
> it requires rustc 1.59 or newer, while the currently active rustc
> version is 1.58.1

[1]: https://github.com/Amanieu/parking_lot/pull/309

[2]: https://github.com/rust-lang/rust/releases/tag/1.59.0