Update use of `libc::timespec` to prepare for future libc version
In a future release of the `libc` crate, `libc::timespec` will contain
private padding fields on `*-linux-musl` targets and so the struct will
no longer be able to be created using the literal initialization syntax.

Update the use of `libc::timespec` to create a value by calling
`std::mem::zeroed()` first and then setting the `tv_sec` and `tv_nsec`
fields manually which works with both current versions of `libc` as well
as the future version which contains that change.